services/ai-service/requirements.txt
flask
numpy
pillow
requests
google-generativeai==0.8.3
protobuf==4.25.3
grpcio==1.62.1